chore: pull CTA Card into separate component
