+++ nextjs-app/app/globals.css
@@ -9,8 +9,6 @@
   :root {
     --font-sans: var(--font-open-sans);
     --font-heading: var(--font-outfit);
-
-
   }
 
   body {
